docs : 스타일 컴포넌트 파일을 따로 분리하는 작업 완료
